price issue is resolved

+++ db/seeds.rb
@@ -1579,7 +1579,7 @@ This cooking apron is generously cut for full coverage on the both bib and botto
 		if params = tops[filename]
 
 				# pro = Shoppe::Product.new(:name => name, :sku => sku, :description => 'test', :short_description => 'test', :weight => 1.119, :price => 24.99, :cost_price => 8.99, :tax_rate => tax_rate)
-				pro = Shoppe::Product.new(default_params.merge(description: params[:description], tax_rate: tax_rate, cost_price: params[:price].to_i, price: params[:price].to_i, colors: params[:colors], sizes: params[:sizes], sku: params[:sku], weight: params[:weight], name: name))
+				pro = Shoppe::Product.new(default_params.merge(description: params[:description], tax_rate: tax_rate, cost_price: params[:price].to_f, price: params[:price].to_f, colors: params[:colors], sizes: params[:sizes], sku: params[:sku], weight: params[:weight], name: name))
 				pro.product_category = cat1
 				pro.default_image_file = get_file(filename)
 				pro.save!
@@ -1595,7 +1595,7 @@ This cooking apron is generously cut for full coverage on the both bib and botto
 
 		elsif params = sets[filename] 
 				
-				pro = Shoppe::Product.new(default_params.merge(description: params[:description], cost_price: params[:price].to_i, tax_rate: tax_rate, price: params[:price].to_i, colors: params[:colors], sizes: params[:sizes], sku: params[:sku], weight: params[:weight], name: name))
+				pro = Shoppe::Product.new(default_params.merge(description: params[:description], cost_price: params[:price].to_f, tax_rate: tax_rate, price: params[:price].to_f, colors: params[:colors], sizes: params[:sizes], sku: params[:sku], weight: params[:weight], name: name))
 				pro.product_category = cat2
 				pro.default_image_file = get_file(filename)
 				pro.save!
@@ -1611,7 +1611,7 @@ This cooking apron is generously cut for full coverage on the both bib and botto
 
 		elsif params = labjackets[filename] 
 				
-				pro = Shoppe::Product.new(default_params.merge(description: params[:description], cost_price: params[:price].to_i, tax_rate: tax_rate, price: params[:price].to_i, colors: params[:colors], sizes: params[:sizes], sku: params[:sku], weight: params[:weight], name: name))
+				pro = Shoppe::Product.new(default_params.merge(description: params[:description], cost_price: params[:price].to_f, tax_rate: tax_rate, price: params[:price].to_f, colors: params[:colors], sizes: params[:sizes], sku: params[:sku], weight: params[:weight], name: name))
 				pro.product_category = cat4
 				pro.default_image_file = get_file(filename)
 				pro.save!
@@ -1627,7 +1627,7 @@ This cooking apron is generously cut for full coverage on the both bib and botto
 
 		elsif params = dresses[filename] 
 
-				pro = Shoppe::Product.new(default_params.merge(description: params[:description], cost_price: params[:price].to_i, tax_rate: tax_rate, price: params[:price].to_i, colors: params[:colors], sizes: params[:sizes], sku: params[:sku], weight: params[:weight], name: name))
+				pro = Shoppe::Product.new(default_params.merge(description: params[:description], cost_price: params[:price].to_f, tax_rate: tax_rate, price: params[:price].to_f, colors: params[:colors], sizes: params[:sizes], sku: params[:sku], weight: params[:weight], name: name))
 				pro.product_category = cat23
 				pro.default_image_file = get_file(filename)
 				pro.save!
@@ -1643,7 +1643,7 @@ This cooking apron is generously cut for full coverage on the both bib and botto
 
 		elsif params = clearance[filename] 
 
-				pro = Shoppe::Product.new(default_params.merge(description: params[:description], cost_price: params[:price].to_i, tax_rate: tax_rate, price: params[:price].to_i, colors: params[:colors], sizes: params[:sizes], sku: params[:sku], weight: params[:weight], name: name))
+				pro = Shoppe::Product.new(default_params.merge(description: params[:description], cost_price: params[:price].to_f, tax_rate: tax_rate, price: params[:price].to_f, colors: params[:colors], sizes: params[:sizes], sku: params[:sku], weight: params[:weight], name: name))
 				pro.product_category = cat22
 				pro.default_image_file = get_file(filename)
 				pro.save!
@@ -1659,7 +1659,7 @@ This cooking apron is generously cut for full coverage on the both bib and botto
 
 		elsif params = stethoscopes[filename] 
 
-				pro = Shoppe::Product.new(default_params.merge(description: params[:description], cost_price: params[:price].to_i, tax_rate: tax_rate, price: params[:price].to_i, colors: params[:colors], sizes: params[:sizes], sku: params[:sku], weight: params[:weight], name: name))
+				pro = Shoppe::Product.new(default_params.merge(description: params[:description], cost_price: params[:price].to_f, tax_rate: tax_rate, price: params[:price].to_f, colors: params[:colors], sizes: params[:sizes], sku: params[:sku], weight: params[:weight], name: name))
 				pro.product_category = cat7
 				pro.default_image_file = get_file(filename)
 				pro.save!
@@ -1675,7 +1675,7 @@ This cooking apron is generously cut for full coverage on the both bib and botto
 
 		elsif params = workcoats[filename] 
 
-				pro = Shoppe::Product.new(default_params.merge(description: params[:description], cost_price: params[:price].to_i, tax_rate: tax_rate, price: params[:price].to_i, colors: params[:colors], sizes: params[:sizes], sku: params[:sku], weight: params[:weight], name: name))
+				pro = Shoppe::Product.new(default_params.merge(description: params[:description], cost_price: params[:price].to_f, tax_rate: tax_rate, price: params[:price].to_f, colors: params[:colors], sizes: params[:sizes], sku: params[:sku], weight: params[:weight], name: name))
 				pro.product_category = cat15
 				pro.default_image_file = get_file(filename)
 				pro.save!
@@ -1691,7 +1691,7 @@ This cooking apron is generously cut for full coverage on the both bib and botto
 
 		elsif params = workco[filename] 
 
-				pro = Shoppe::Product.new(default_params.merge(description: params[:description], cost_price: params[:price].to_i, tax_rate: tax_rate, price: params[:price].to_i, colors: params[:colors], sizes: params[:sizes], sku: params[:sku], weight: params[:weight], name: name))
+				pro = Shoppe::Product.new(default_params.merge(description: params[:description], cost_price: params[:price].to_f, tax_rate: tax_rate, price: params[:price].to_f, colors: params[:colors], sizes: params[:sizes], sku: params[:sku], weight: params[:weight], name: name))
 				pro.product_category = cat16
 				pro.default_image_file = get_file(filename)
 				pro.save!
@@ -1707,7 +1707,7 @@ This cooking apron is generously cut for full coverage on the both bib and botto
 
 		elsif params = fireretardent[filename] 
 
-				pro = Shoppe::Product.new(default_params.merge(description: params[:description], cost_price: params[:price].to_i, tax_rate: tax_rate, price: params[:price].to_i, colors: params[:colors], sizes: params[:sizes], sku: params[:sku], weight: params[:weight], name: name))
+				pro = Shoppe::Product.new(default_params.merge(description: params[:description], cost_price: params[:price].to_f, tax_rate: tax_rate, price: params[:price].to_f, colors: params[:colors], sizes: params[:sizes], sku: params[:sku], weight: params[:weight], name: name))
 				pro.product_category = cat14
 				pro.default_image_file = get_file(filename)
 				pro.save!
@@ -1723,7 +1723,7 @@ This cooking apron is generously cut for full coverage on the both bib and botto
 
 		elsif params = workshirts[filename] 
 
-				pro = Shoppe::Product.new(default_params.merge(description: params[:description], cost_price: params[:price].to_i, tax_rate: tax_rate, price: params[:price].to_i, colors: params[:colors], sizes: params[:sizes], sku: params[:sku], weight: params[:weight], name: name))
+				pro = Shoppe::Product.new(default_params.merge(description: params[:description], cost_price: params[:price].to_f, tax_rate: tax_rate, price: params[:price].to_f, colors: params[:colors], sizes: params[:sizes], sku: params[:sku], weight: params[:weight], name: name))
 				pro.product_category = cat18
 				pro.default_image_file = get_file(filename)
 				pro.save!
@@ -1739,7 +1739,7 @@ This cooking apron is generously cut for full coverage on the both bib and botto
 
 		elsif params = workvests[filename] 
 
-				pro = Shoppe::Product.new(default_params.merge(description: params[:description], cost_price: params[:price].to_i, tax_rate: tax_rate, price: params[:price].to_i, colors: params[:colors], sizes: params[:sizes], sku: params[:sku], weight: params[:weight], name: name))
+				pro = Shoppe::Product.new(default_params.merge(description: params[:description], cost_price: params[:price].to_f, tax_rate: tax_rate, price: params[:price].to_f, colors: params[:colors], sizes: params[:sizes], sku: params[:sku], weight: params[:weight], name: name))
 				pro.product_category = cat19
 				pro.default_image_file = get_file(filename)
 				pro.save!
@@ -1755,7 +1755,7 @@ This cooking apron is generously cut for full coverage on the both bib and botto
 
 		elsif params = wvest[filename] 
 
-				pro = Shoppe::Product.new(default_params.merge(description: params[:description], cost_price: params[:price].to_i, tax_rate: tax_rate, price: params[:price].to_i, colors: params[:colors], sizes: params[:sizes], sku: params[:sku], weight: params[:weight], name: name))
+				pro = Shoppe::Product.new(default_params.merge(description: params[:description], cost_price: params[:price].to_f, tax_rate: tax_rate, price: params[:price].to_f, colors: params[:colors], sizes: params[:sizes], sku: params[:sku], weight: params[:weight], name: name))
 				pro.product_category = cat13
 				pro.default_image_file = get_file(filename)
 				pro.save!
@@ -1771,7 +1771,7 @@ This cooking apron is generously cut for full coverage on the both bib and botto
 
 		elsif params = caps[filename] 
 
-				pro = Shoppe::Product.new(default_params.merge(description: params[:description], cost_price: params[:price].to_i, tax_rate: tax_rate, price: params[:price].to_i, colors: params[:colors], sizes: params[:sizes], sku: params[:sku], weight: params[:weight], name: name))
+				pro = Shoppe::Product.new(default_params.merge(description: params[:description], cost_price: params[:price].to_f, tax_rate: tax_rate, price: params[:price].to_f, colors: params[:colors], sizes: params[:sizes], sku: params[:sku], weight: params[:weight], name: name))
 				pro.product_category = cat5
 				pro.default_image_file = get_file(filename)
 				pro.save!
@@ -1787,7 +1787,7 @@ This cooking apron is generously cut for full coverage on the both bib and botto
 
 		elsif params = accessories[filename] 
 
-				pro = Shoppe::Product.new(default_params.merge(description: params[:description], cost_price: params[:price].to_i, tax_rate: tax_rate, price: params[:price].to_i, colors: params[:colors], sizes: params[:sizes], sku: params[:sku], weight: params[:weight], name: name))
+				pro = Shoppe::Product.new(default_params.merge(description: params[:description], cost_price: params[:price].to_f, tax_rate: tax_rate, price: params[:price].to_f, colors: params[:colors], sizes: params[:sizes], sku: params[:sku], weight: params[:weight], name: name))
 				pro.product_category = cat6
 				pro.default_image_file = get_file(filename)
 				pro.save!
@@ -1803,7 +1803,7 @@ This cooking apron is generously cut for full coverage on the both bib and botto
 
 		elsif params = workpants[filename] 
 
-				pro = Shoppe::Product.new(default_params.merge(description: params[:description], cost_price: params[:price].to_i, tax_rate: tax_rate, price: params[:price].to_i, colors: params[:colors], sizes: params[:sizes], sku: params[:sku], weight: params[:weight], name: name))
+				pro = Shoppe::Product.new(default_params.merge(description: params[:description], cost_price: params[:price].to_f, tax_rate: tax_rate, price: params[:price].to_f, colors: params[:colors], sizes: params[:sizes], sku: params[:sku], weight: params[:weight], name: name))
 				pro.product_category = cat17
 				pro.default_image_file = get_file(filename)
 				pro.save!
@@ -1819,7 +1819,7 @@ This cooking apron is generously cut for full coverage on the both bib and botto
 
 		elsif params = pants[filename] 
 
-				pro = Shoppe::Product.new(default_params.merge(description: params[:description], cost_price: params[:price].to_i, tax_rate: tax_rate, price: params[:price].to_i, colors: params[:colors], sizes: params[:sizes], sku: params[:sku], weight: params[:weight], name: name))
+				pro = Shoppe::Product.new(default_params.merge(description: params[:description], cost_price: params[:price].to_f, tax_rate: tax_rate, price: params[:price].to_f, colors: params[:colors], sizes: params[:sizes], sku: params[:sku], weight: params[:weight], name: name))
 				pro.product_category = cat3
 				pro.default_image_file = get_file(filename)
 				pro.save!
@@ -1835,7 +1835,7 @@ This cooking apron is generously cut for full coverage on the both bib and botto
 
 		elsif params = aprons[filename] 
 
-				pro = Shoppe::Product.new(default_params.merge(description: params[:description], cost_price: params[:price].to_i, tax_rate: tax_rate, price: params[:price].to_i, colors: params[:colors], sizes: params[:sizes], sku: params[:sku], weight: params[:weight], name: name))
+				pro = Shoppe::Product.new(default_params.merge(description: params[:description], cost_price: params[:price].to_f, tax_rate: tax_rate, price: params[:price].to_f, colors: params[:colors], sizes: params[:sizes], sku: params[:sku], weight: params[:weight], name: name))
 				pro.product_category = cat12
 				pro.default_image_file = get_file(filename)
 				pro.save!
@@ -1851,7 +1851,7 @@ This cooking apron is generously cut for full coverage on the both bib and botto
 
 		elsif params = chefpants[filename] 
 
-				pro = Shoppe::Product.new(default_params.merge(description: params[:description], cost_price: params[:price].to_i, tax_rate: tax_rate, price: params[:price].to_i, colors: params[:colors], sizes: params[:sizes], sku: params[:sku], weight: params[:weight], name: name))
+				pro = Shoppe::Product.new(default_params.merge(description: params[:description], cost_price: params[:price].to_f, tax_rate: tax_rate, price: params[:price].to_f, colors: params[:colors], sizes: params[:sizes], sku: params[:sku], weight: params[:weight], name: name))
 				pro.product_category = cat10
 				pro.default_image_file = get_file(filename)
 				pro.save!
@@ -1867,7 +1867,7 @@ This cooking apron is generously cut for full coverage on the both bib and botto
 
 		elsif params = chefcoats[filename] 
 
-				pro = Shoppe::Product.new(default_params.merge(description: params[:description], cost_price: params[:price].to_i, tax_rate: tax_rate, price: params[:price].to_i, colors: params[:colors], sizes: params[:sizes], sku: params[:sku], weight: params[:weight], name: name))
+				pro = Shoppe::Product.new(default_params.merge(description: params[:description], cost_price: params[:price].to_f, tax_rate: tax_rate, price: params[:price].to_f, colors: params[:colors], sizes: params[:sizes], sku: params[:sku], weight: params[:weight], name: name))
 				pro.product_category = cat8
 				pro.default_image_file = get_file(filename)
 				pro.save!
@@ -1883,7 +1883,7 @@ This cooking apron is generously cut for full coverage on the both bib and botto
 
 		elsif params = cheftops[filename] 
 
-				pro = Shoppe::Product.new(default_params.merge(description: params[:description], cost_price: params[:price].to_i, tax_rate: tax_rate, price: params[:price].to_i, colors: params[:colors], sizes: params[:sizes], sku: params[:sku], weight: params[:weight], name: name))
+				pro = Shoppe::Product.new(default_params.merge(description: params[:description], cost_price: params[:price].to_f, tax_rate: tax_rate, price: params[:price].to_f, colors: params[:colors], sizes: params[:sizes], sku: params[:sku], weight: params[:weight], name: name))
 				pro.product_category = cat11
 				pro.default_image_file = get_file(filename)
 				pro.save!
@@ -1899,7 +1899,7 @@ This cooking apron is generously cut for full coverage on the both bib and botto
 
 		elsif params = chefhats[filename] 
 
-				pro = Shoppe::Product.new(default_params.merge(description: params[:description], cost_price: params[:price].to_i, tax_rate: tax_rate, price: params[:price].to_i, colors: params[:colors], sizes: params[:sizes], sku: params[:sku], weight: params[:weight], name: name))
+				pro = Shoppe::Product.new(default_params.merge(description: params[:description], cost_price: params[:price].to_f, tax_rate: tax_rate, price: params[:price].to_f, colors: params[:colors], sizes: params[:sizes], sku: params[:sku], weight: params[:weight], name: name))
 				pro.product_category = cat9
 				pro.default_image_file = get_file(filename)
 				pro.save!
@@ -1915,7 +1915,7 @@ This cooking apron is generously cut for full coverage on the both bib and botto
 
 		elsif params = mentalityh[filename] 
 
-				pro = Shoppe::Product.new(default_params.merge(description: params[:description], cost_price: params[:price].to_i, tax_rate: tax_rate, price: params[:price].to_i, colors: params[:colors], sizes: params[:sizes], sku: params[:sku], weight: params[:weight], name: name))
+				pro = Shoppe::Product.new(default_params.merge(description: params[:description], cost_price: params[:price].to_f, tax_rate: tax_rate, price: params[:price].to_f, colors: params[:colors], sizes: params[:sizes], sku: params[:sku], weight: params[:weight], name: name))
 				pro.product_category = cat20
 				pro.default_image_file = get_file(filename)
 				pro.save!
@@ -1931,7 +1931,7 @@ This cooking apron is generously cut for full coverage on the both bib and botto
 
 		elsif params = mentalitysf[filename] 
 
-		pro = Shoppe::Product.new(default_params.merge(description: params[:description], cost_price: params[:price].to_i, tax_rate: tax_rate, price: params[:price].to_i, colors: params[:colors], sizes: params[:sizes], sku: params[:sku], weight: params[:weight], name: name))
+		pro = Shoppe::Product.new(default_params.merge(description: params[:description], cost_price: params[:price].to_f, tax_rate: tax_rate, price: params[:price].to_f, colors: params[:colors], sizes: params[:sizes], sku: params[:sku], weight: params[:weight], name: name))
 				pro.product_category = cat21
 				pro.default_image_file = get_file(filename)
 				pro.save!
